modify layout and fix quiz topics data fetching
